merge

+++ src/ast/usage.h
@@ -10,7 +10,7 @@ namespace luisa::compute {
 
 /** @file */
 /// Flags of usage
-enum struct Usage : uint8_t {
+enum struct Usage : uint32_t {
     NONE = 0u,
     READ = 0x01u,
     WRITE = 0x02u,
